requirements.txt
Django==5.0.1
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
openstacksdk==1.5.0
redfish>=3.1.0
django-auth-ldap==4.6.0
python-ldap==3.4.4
django-jazzmin==2.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
PyYAML==6.0.1
